+++ src/StarSystem.cpp
@@ -532,8 +532,8 @@ void StarSystem::MakeBinaryPair(SBody *a, SBody *b, fixed minDist, MTRand &rand)
 	a->orbit.semiMajorAxis = AU * (a->semiMajorAxis * a0).ToDouble();
 	a->orbit.period = 60*60*24*365* a->semiMajorAxis.ToDouble() * sqrt(a->semiMajorAxis.ToDouble() / m.ToDouble());
 	
-	const float rotY = rand.Double()*M_PI/2.0;
-	const float rotZ = rand.Double(M_PI);
+	const float rotY = (float)(rand.Double()*M_PI/2.0);
+	const float rotZ = (float)rand.Double(M_PI);
 	a->orbit.rotMatrix = matrix4x4d::RotateYMatrix(rotY) * matrix4x4d::RotateZMatrix(rotZ);
 	b->orbit.rotMatrix = matrix4x4d::RotateYMatrix(rotY) * matrix4x4d::RotateZMatrix(rotZ-M_PI);
